added frontend for product api

+++ Backend/src/utils/dynamoDb/productManagement.ts
@@ -1,6 +1,6 @@
-import { ProductInfo } from '@models/index';
+import { ProductInfo, ConfigItem, PlatformInfo, TestSuite, DeviceInfo } from '@models/index';
 import * as TableNames from '@utils/dynamoDb/getTableNames';
-import { appLogger} from '@utils/index';
+import { appLogger, getProductConfig, getPlatformsList, getTestSuites, getDevicesList} from '@utils/index';
 import { DynamoDB } from 'aws-sdk';
 //import { getUserDocumentFromEmail } from './getUserDocument';
 import { deleteItem, get, put, scan, update } from './sdk';
@@ -13,6 +13,8 @@ export const createProduct = async (productData: ProductInfo, userId: string): P
       appLogger.error(err);
       throw err;
     }
+    console.log(productData, "productData");
+    console.log(productData.devices, "productData.devices");
     const item: ProductInfo = {
         devices: productData.devices,
         id: productData.id,
@@ -24,15 +26,59 @@ export const createProduct = async (productData: ProductInfo, userId: string): P
         testSuite: productData.testSuite,
         version: productData.version
       };
+
+    Object.keys(productData).forEach((val, i) => {
+      if (
+        !(
+          val === 'id' ||
+          val === 'name' ||
+          val === 'version'
+        )
+      ) {
+        if (productData[val].length > 0 && typeof productData[val] === 'object') {
+          item[val] = new Array();
+          productData[val].forEach((ele: any, j: number) => {
+            if ((typeof(ele) === 'string') && (ele.match(regex))) {
+                item[val].push(ele.split('Other:')[1]);
+            } else {
+              item[val].push(ele);
+            }
+          });
+        } else {
+          item[val] = productData[val];
+        }
+      }
+    });
+
     const params: DynamoDB.PutItemInput = <DynamoDB.PutItemInput>(<unknown>{
         ConditionExpression: 'attribute_not_exists(id)',
         Item: item,
-        TableName: TableNames.getCampaignsTableName(),
+        TableName: TableNames.getProductsTableName(),
       });
     appLogger.info({ createCampaign_put_params: params });
     return put<ProductInfo>(params);
 }
+export const getCreateProductConfig = async (
+  orgId: string
+): Promise<ConfigItem> => {
+  const productConfig: ConfigItem = await getProductConfig(orgId);
+  appLogger.info({ getProductConfig: productConfig });
 
+  const platforms: PlatformInfo[] = await getPlatformsList();
+  const platforms_key = 'platforms';
+  productConfig.config[platforms_key].options = {};
+  platforms.forEach((val: PlatformInfo) => productConfig.config[platforms_key].options[val.id] = val.name);
+  const testSuite: TestSuite[] = await getTestSuites();
+  const testSuite_key = 'testSuite';
+  productConfig.config[testSuite_key].options = {};
+  testSuite.forEach((val: PlatformInfo) => productConfig.config[testSuite_key].options[val.id] = val.name);
+  const devices: DeviceInfo[] = await getDevicesList();
+  const devices_key = 'devices';
+  productConfig.config[devices_key].options = {};
+  devices.forEach((val: DeviceInfo) => productConfig.config[devices_key].options[val.id] = val.name);
+
+  return productConfig;
+};
 export const getProductDetails = async (id: string,version: string): Promise<ProductInfo> => {
     const params: DynamoDB.GetItemInput = <DynamoDB.GetItemInput>(<unknown>{
       Key: {
@@ -54,15 +100,21 @@ appLogger.info({ getProductsList_scan_params: params });
 return scan<ProductInfo[]>(params);
 };
 
-export const deleteproduct = async (id: string): Promise<ProductInfo | undefined> => {
-const params: DynamoDB.DeleteItemInput = <DynamoDB.DeleteItemInput>(<unknown>{
-    Key: {
-    id,
-    },
-    TableName: TableNames.getProductsTableName(),
-});
-appLogger.info({ deleteProduct_delete_params: params });
-return deleteItem(params);
+export const deleteProduct = async (id: string, version: string, userId: string): Promise<ProductInfo | undefined> => {
+  if (!userId) {
+    const err = new Error('Unauthorized attempt');
+    appLogger.error(err);
+    throw err;
+  }
+  const params: DynamoDB.DeleteItemInput = <DynamoDB.DeleteItemInput>(<unknown>{
+      Key: {
+      id,
+      version,
+      },
+      TableName: TableNames.getProductsTableName(),
+  });
+  appLogger.info({ deleteProduct_delete_params: params });
+  return deleteItem(params);
 };
 
 export const updateProduct = async (updateInfo: ProductInfo, userId: string) => {
@@ -80,7 +132,7 @@ export const updateProduct = async (updateInfo: ProductInfo, userId: string) => 
     Object.keys(updateInfo).forEach((val, i) => {
       if (
         !(
-          val === 'id'
+          val === 'id' || val === 'version'
         )
       ) {
         if (
@@ -113,6 +165,7 @@ export const updateProduct = async (updateInfo: ProductInfo, userId: string) => 
       ExpressionAttributeValues: EAV,
       Key: {
         id: updateInfo.id,
+        version: updateInfo.version,
       },
       TableName: TableNames.getProductsTableName(),
       UpdateExpression: SET,
